atividade_ocorrencias_aereas/requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
plotly>=5.22.0
numpy>=1.26.4
pillow>=10.3.0
